+++ discord/enums.py
@@ -79,6 +79,7 @@ __all__ = (
     'ScheduledEventStatus',
     'Locale',
     'MFALevel',
+    'InputTextStyle',
 )
 
 if TYPE_CHECKING:
@@ -566,6 +567,7 @@ class InteractionType(Enum):
     ping = 1
     application_command = 2
     component = 3
+    modal_submit = 5
 
 
 class InteractionResponseType(Enum):
@@ -577,7 +579,7 @@ class InteractionResponseType(Enum):
     deferred_message_update = 6  # for components
     message_update = 7  # for components
     application_command_autocomplete_result = 8  # for autocomplete application command options
-
+    modal = 9 # for modal interactions
 
 class VideoQualityMode(Enum):
     auto = 1
@@ -591,6 +593,7 @@ class ComponentType(Enum):
     action_row = 1
     button = 2
     select = 3
+    input_text = 4
 
     def __int__(self) -> int:
         return self.value
@@ -713,6 +716,11 @@ class MFALevel(Enum):
 
     def __int__(self) -> int:
         return self.value
+
+
+class InputTextStyle(Enum):
+    short = 1  # for short single-line text
+    paragraph = 2  # for much longer inputs
 
 
 T = TypeVar('T')