+++ discord/state.py
@@ -72,6 +72,7 @@ from .invite import Invite
 from .integrations import _integration_factory
 from .interactions import Interaction
 from .ui.view import ViewStore, View
+from .ui.modal import ModalStore, Modal
 from .stage_instance import StageInstance
 from .threads import Thread, ThreadMember
 from .sticker import GuildSticker
@@ -294,7 +295,7 @@ class ConnectionState:
 
         self.clear()
 
-    def clear(self, *, views: bool = True) -> None:
+    def clear(self, *, views: bool = True, modals: bool = True) -> None:
         self.user: Optional[ClientUser] = None
         # Originally, this code used WeakValueDictionary to maintain references to the
         # global user mapping.
@@ -314,6 +315,8 @@ class ConnectionState:
         self._guilds: Dict[int, Guild] = {}
         if views:
             self._view_store: ViewStore = ViewStore(self)
+        if modals:
+            self._modal_store: ModalStore = ModalStore(self)
 
         self._voice_clients: Dict[int, VoiceProtocol] = {}
 
@@ -431,6 +434,9 @@ class ConnectionState:
     @property
     def guilds(self) -> List[Guild]:
         return list(self._guilds.values())
+
+    def store_modal(self, modal: Modal, user_id: int) -> None:
+        self._modal_store.add_modal(modal, user_id)
 
     def _get_guild(self, guild_id: Optional[int]) -> Optional[Guild]:
         if guild_id is None:
@@ -635,7 +641,7 @@ class ConnectionState:
             self._ready_task.cancel()
 
         self._ready_state = asyncio.Queue()
-        self.clear(views=False)
+        self.clear(views=False, modals=False)
         self.user = ClientUser(state=self, data=data['user'])
         self.store_user(data['user'])
 
@@ -881,6 +887,9 @@ class ConnectionState:
                 asyncio.create_task(command_option._define_autocomplete_result(response=response, command=used_command))  # type: ignore
 
                 break
+        elif data['type'] == 5:  # modal submit
+            custom_id = interaction.data['custom_id']  # type: ignore
+            self._modal_store.dispatch(custom_id, interaction)
 
         self.dispatch('interaction', interaction)